requirements.txt
torch
matplotlib
numpy
glob2
scikit-image
scipy
Pillow